refactor: convert rerxnplus.js to TypeScript with prop types validation
